[HSA-239 Fix]make the valid test case and invalid test case run correctly

+++ include/brig_util.h
@@ -106,7 +106,7 @@ template<class Super> class brig_iterator : private Super {
     return this->curr != other.curr;
   }
 
-  Self operator+(intptr_t addend) const {
+  Self operator+(uint64_t addend) const {
     brig_iterator other = *this;
     for(unsigned i = 0; i < addend; ++i) {
       ++other;
